Rename previously implemented tasks to Worker Task
