+++ LDLSH/src/main/java/SystemLayer/Components/TaskImpl/Worker/WorkerTask.java
@@ -1,0 +1,9 @@
+package SystemLayer.Components.TaskImpl.Worker;
+
+import SystemLayer.Data.DataObjectsImpl.DataObject;
+
+import java.util.concurrent.Callable;
+
+public interface WorkerTask extends Callable<DataObject> {
+
+}
